Don't shuffle FGR loaded from Project64
I couldn't receive any comments from zilmar about the way the FGR
register is stored inside the savegame of pj64. So I wrote some
savegames to help me to reverse engineer it. The result was that we do
it nearly perfect, but the PJ64 savegame doesn't shuffle the data back
to 64 bit wide FGR before saving when it is currently in 32 bit wide
FGR. So we shouldn't do it when saving to or loading from PJ64 savegame.