Don't override FGR with 0xffffffff after mtc0
When somebody switches between 64-bit wide FGR to 32-bit FGR it will
rearrange all registers in memory. This is partly done by shifting the
different parts of a new register pair for a double together. It is
important that this is done with unsigned integers or otherwise sign
extension is done and the upper part of the register will always be
0xffffffff when the bit 31 of the lower part of the register is 1.