Prevent over-optimization of rjump related data
GCC tries to remove and/or rename local symbols when trying to optimize
the code. This is a common problem when using link time optimization or
other optimization with deep inspection of dataflow. Inline assembler
code has to be annotated when it tries to access data behind such
symbols to prevent over optimization and to correctly name these
symbols.

This also removes partially the requirement of having different
implementation of the rjump code for different combination of OS and
PIC/PIE targets.