Prevent data corruption with caller managed OSD text
OSD text like the volume information is update by the caller in main.c, but the
drawing function in osd.c maybe already deleted the corresponding data
structure. The workaround was to check if the pointer to data structure could
be found in the list. It could happen that another text was shown inbetween and
the libc reused the old memory region for another text message. The
osd_message_valid function would wrongly inform the caller that the OSD text
wasn't deleted and the following osd_update_message would modify the wrong
message.

We must forbid that the messages is deleted by both the caller and the osd
drawing mainloop. The mainloop restricts itself to removing it from the list
when the caller marked it as user_managed directly after it was allocated. The
caller has to remove it manually using osd_delete_message in that situation.