Use test rom from separate repository mupen64plus-test
The test rom has an own repository called mupen64plus-test and doesn't have to
be shipped in tarballs for mupen64plus-core. This also makes the
mupen64plus-core-norom module tarball obsolete because the rom is never
included. The rom is still included in the source and build bundle versions of
mupen64plus.