Don't store mapped host registers on stack before dynarec call
The dynamic recompiler uses a simple register mapping for currently used
registers to reduce memory access for operations. The content of these
registers have to be saved back to the memory region containing the emulated
register bar before they will be remapped or may get damaged by an unsave
operation.

Emulated operations in the amd64 dynamic recompiler tried to avoid the complex
register free'ing operation before an call to a C function by storing the
content of mapped registers to the stack. This implementation is faulty because

 * the return address (top most stack position when the dynamic recompiler
   started a recompiled block) may get written during the execution of the C
   function. The problem is that the return address is moved to another
   position when the registers were stored in the stack and therefore the
   content of a saved register gets damaged by this write.
 * the called C function may want to access the content of an emulated
   register. This is done by reading the content of the emulated register bar
   in memory. This read will return an outdated value because the new value is
   only stored on the stack.

This problem can be avoided without more complicated checks by not saving the
content on the stack and instead unmapping the registers and saving their
content in the emulated register bar.