Allow to crosscompile and compile extra modules with m64p_helper_scripts
The install targets of the Makefiles are now used to generate the test
directory. This allows the build script to use the correct file extension and
similar differences between the platforms.

A user with the correct mercurial clones in sources/ can also use the
environment variable M64P_COMPONENTS to specify which component should be
compiled. This is interesting when also wanting to compile/update wahrhafts
plugins in one run.