debugger: Split opcodes from their operands on tabs as well as spaces
For some reason, Coprocessor 1 opcodes are written by the disassembler
as

  opcode <tab> operands

while other opcodes use a 16-character field for the opcode, padded
with spaces. r4300_decode_op would split Coprocessor 1 opcodes like so:

  opcode "mtc1\t\t$a0,$f0", operands ""

Now they are split like so:

  opcode "mtc1", operands "$a0,$f0"

This can be helpful for disassembly outputs that don't use tabs.