Change sign_extended to act like an expression instead of a fixup
Previous code to sign-extend a 32-bit result into 64 bits looked like

  rrt32 = ...;
  sign_extended(rrt);

while it is now

  rrt = SE32(...);

This change may appear to be insignificant, but the C code now has one
less store in it, and it may perform the sign-extension on a temporary
variable instead of needing to reload due to the 64-bit version of the
variable aliasing with parts of the 32-bit version.

On amd64, using gcc 4.8.2, the code for ADDIU with sign_extended as a
fixup looked like this:

48 8d 05 71 14 e9 00    lea    0xe91471(%rip),%rax   # &PC
48 8b 10                mov    (%rax),%rdx           # PC
48 8b 72 08             mov    0x8(%rdx),%rsi        # PC->f.i.rs
48 8b 4a 10             mov    0x10(%rdx),%rcx       # PC->f.i.rt
0f bf 52 18             movswl 0x18(%rdx),%edx       # (int32_t) PC->f.i.immediate
03 16                   add    (%rsi),%edx           # *rs + imm
89 11                   mov    %edx,(%rcx)           # store to *rt (32-bit)
48 8b 10                mov    (%rax),%rdx           # RELOAD PC
48 8b 52 10             mov    0x10(%rdx),%rdx       # RELOAD PC->f.i.rt
48 63 0a                movslq (%rdx),%rcx           # RELOAD *rt, sign-extend
48 89 0a                mov    %rcx,(%rdx)           # store to *rt (64-bit)
48 81 00 d0 00 00 00    addq   $0xd0,(%rax)          # PC++;
c3                      retq

This store and reload can be serviced on some chips by store-forwarding
inside the data cache, but Intel Atom chips, for example, cannot handle
this very well.

With SE32 as an expression, ADDIU looks like this instead:

48 8d 05 71 14 e9 00    lea    0xe91471(%rip),%rax
48 8b 10                mov    (%rax),%rdx
48 8b 72 08             mov    0x8(%rdx),%rsi
48 8b 4a 10             mov    0x10(%rdx),%rcx
0f bf 52 18             movswl 0x18(%rdx),%edx
03 16                   add    (%rsi),%edx
48 63 d2                movslq %edx,%rdx
48 89 11                mov    %rdx,(%rcx)           # one store (64-bit)
48 81 00 d0 00 00 00    addq   $0xd0,(%rax)
c3                      retq