Unify global variables, part 1: Interpreters and Hacktarux JIT
The following global variables are now in a single contiguous structure
that is aligned to a page (4096 bytes) in compilers that support it:

reg[32], hi, lo, g_cp0_regs[32], FCR0, FCR31, llbit,
reg_cop1_simple[32], reg_cop1_double[32], reg_cop1_fgr_64[32],
delay_slot, next_interupt, address, rdword, cpu_byte, cpu_hword,
cpu_word, cpu_dword

On many architectures, the structure allows access to this data with an
offset from a base register containing the address of the structure. It
being aligned to a page allows it to require only one TLB entry, if the
compiler supports it.

Notably missing from this commit are the modifications required for the
New Dynarec to work with these new structure offsets.