Unify global variables, part 2: New Dynarec/ARM
Much like before, there is conditional compilation to selectively have
the New Dynarec create a variable in its own memory space instead of C
putting it somewhere inconvenient.

Now, though, the New Dynarec is just adopting the contiguous structure
that is also used by the other cores, so it is no longer the only core
that has the advantage of a contiguous set of often-used variables.