Trace JIT for MIPS32 and AMD64 - EXPERIMENTAL
EXPERIMENTAL CODE WARNING: This commit is extremely dirty and is not
meant to be used in production code yet. It hardcodes the use of the
Trace JIT, as well as the exception and interrupt handling code that
it requires for efficient code. It is also one big commit which does
not properly describe the work that went into it, and it lacks docs.