Move some mem variables into r4300 module to ease information passing.
These variables (address, wbyte, whword, wword, wdword, rdword) were
originally inside the memory module because they are memory related. But
if we look closer, these variables are just function arguments for
parameterless functions, which is a limitation imposed by the dynarec.
So migrating to the r4300_core module will help isolate this limitation
where it belongs.