Fix undefined behavior in the interpreter
Partially fixes #87, but there are probably more potential overflows
lurking in there; I didn't go over it too thoroughly.

FYI: In C, integer overflow when adding and multiplying signed integers
is undefined, as is shifting any integer by too high a width, and
left-shifting a signed integer if it changes the sign.  GCC started
performing optimizations based on these a decade ago, but those
optimizations don't usually apply in any particular situation, so
usually you just get away with it.

But on my system, with clang-802.0.42, Mario 64 actually fails to boot
in the interpreter without this patch: it fails the ROM checksum test
performed early on (at 0x800001AC).  (Not sure exactly which operation
was responsible; I compiled with -fsanitize=undefined, fixed everything
that made it crash under that and some nearby code, and now Mario 64
works.  However, all of the changes here should be correct/necessary.)