Stubbing mouse controller implementation.
As this is fairly similar to a standard controller (at the pif
protocol level), I just implemented it using a game controller "flavor"
and not a separate entity.

Still WIP and not accessible (unless you hack your way inside main.c).

Also, in the future it should be possible to implement other kind of
"simple" game controllers such as the train controller.