Bump savestate format to 1.2
Recent development omitted to save several emulated device state
variables. We remedy this by appending them at the end of the savestate
file. When loading older savestate files (<1.2 or pj64) we adopt a best-effort
approach to set the missing state to a reasonnable state. But it may
cause some instability.