Avoid warning about casting void* -> function pointer.
This conversion is undefined in strict C ISO and a conforming compiler
must issue a warning about it.  However it is largely supported (and
even possibly mandated for compilers supporting the XSI extension, given
the dlsym prototype). The recommended way of avoiding the warning is
instead to use the *(void**)(&func) trick.

See http://pubs.opengroup.org/onlinepubs/009695399/functions/dlsym.html