Add SiDmaDuration option.
This is a hack on SI DMA timings to allow games with different timing
requirements to work.

We keep the previously hardcoded value of 0x900 as the default value
but fine-tune it for Tetris 64 (J) which can be made to boot with a
lower value of it.

Therefore we can get both Body Harvest and Tertis 64 (J) to work.