safer config file writing
The configuration file writing is subject to two problem situations that
could end up leaving the file empty or with incomplete data:
1. If the program were to terminate while writing to the configuration
   file.
2. If the system were to crash or lose power shortly after the program
   has finished writing the configuration file.

These are covered in detail in:
https://lwn.net/Articles/457667/
https://lwn.net/Articles/322823/

Use a temporary file to mitigate issue #1.

Use fsync() on the file and the directory to mitigate issue #2. This is
not implemented for Windows because I was not able to find evidence that
Windows supports the necessary system calls (or wrappers); I can change
this if I get some feedback giving me direction.