new_dynarec: Save previous register mapping before BNE/BNEL 32bits optim
If one of the source register for BNE/BNEL is r0,
the other source register is considered 32bits if the branch is not taken.

The upper part of this register is thus evicted from the register map.

This was done before the previous register map was saved,
causing the evicted register do be not written back before the branch