Increase maximum allowed overclocking factor to 20
This can be used for particularly demanding game sections when
coupled with 60 FPS cheats to ensure greater framerate stability,
at the cost of increased CPU utilization.

For reference, 1964GEPD supports values up to 18. Even with an overclock
factor of 18, there is still ample room to run above fullspeed on an i9-13900K
on Perfect Dark, so a value of 20 should run well on today's high-end CPUs.